Refactor package.json and yarn.lock for dependency updates
Updated the package.json file to include the latest stable version of the "memoizee" dependency. Also, updated the yarn.lock file to reflect the changes in the dependency tree. This update ensures that the project is using the most up-to-date version of the "memoizee" library, which may include bug fixes and performance improvements.