Refactor show cancellation condition
The condition to check if a show can be cancelled has been refactored. Instead of checking if the price amount is zero, it now checks if there are no outstanding tickets (i.e., reserved and sold tickets minus refunded tickets equals zero). Also, an unnecessary line at the end of the 'show.ts' file was removed.